+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Improved In-Editor parameters rendering [#1443](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1443)
 - Support tabulation for In-Editor parameters [#1444](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1444)
 - Respect an editor for `FlexibleSearch` inlay hints [#1446](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1446)
+- Refresh a parameters panel for all `FlexibleSearch` editors on PSI-elements tree change [#1447](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1447)
 
 ### Fixes
 - Fixed [#1205](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1205) JSoap's inability to parse response with the `&quot;` sequence [#1439](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1439)